Build out the flash component
I thought I could do this with just hooks, but it turns out I did need the context in order to have a single flash area that can be set by components in other parts of the React tree. The context pulls it all together, using the `useContext` hook inside `useFlash`.

I'm definitely not sure if this is the right way to do this, but it works.